requirements.txt
absl-py==1.4.0
accelerate==0.32.1
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
badam==1.2.2
better-exceptions==0.3.3
boltons==23.0.0
boto3==1.34.145
botocore==1.34.145
brotlipy==0.7.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
clip==1.0
cn2an==0.5.22
colorama==0.4.6
conda==23.3.1
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
contourpy==1.0.7
cpm-kernels==1.0.11
cryptography==39.0.1
cycler==0.11.0
datasets==2.19.1
decorator==5.1.1
deepspeed==0.14.4
diffuser==0.0.1
dill==0.3.6
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
einops==0.5.0
email_validator==2.2.0
entrypoints==0.4
evaluate==0.4.2
executing==1.2.0
fairscale==0.4.13
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2024.3.1
ftfy==6.1.1
func-timeout==4.3.5
fuzzywuzzy==0.18.0
gmpy2==2.1.2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gradio==4.38.1
gradio_client==1.1.0
grpcio==1.54.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==2.9
imageio==2.28.0
immutabledict==4.2.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
latex2mathml==3.75.3
lazy_loader==0.2
Levenshtein==0.25.1
lightning-utilities==0.8.0
linkify-it-py==2.0.0
llamafactory==0.8.2.dev0
lmdb==1.4.1
loguru==0.7.0
ltp==4.2.14
ltp-core==0.1.4
ltp-extension==0.1.13
lxml==5.2.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdtex2html==1.2.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmengine-lite==0.10.4
mpi4py==3.1.4
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==2.8.4
ninja==1.11.1
nltk==3.8
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
nvitop==1.3.2
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.35.15
OpenCC==1.1.7
opencompass==0.2.4
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
orjson==3.8.10
packaging==23.0
pandas==1.5.3
parso==0.8.3
peft==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==23.0.1
platformdirs==4.2.2
pluggy==1.0.0
portalocker==2.10.1
prettytable==3.10.2
proces==0.1.7
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.4
pycparser==2.20
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyext==0.7
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypinyin==0.51.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytorch-ignite==0.4.11
pytorch-lightning==2.0.1.post0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rank-bm25==0.2.2
rapidfuzz==3.9.4
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.1
rouge==1.0.1
rouge-chinese==1.0.3
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruff==0.5.3
s3transfer==0.10.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
setuptools==66.0.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sniffio==1.3.0
sse-starlette==2.1.2
stack-data==0.6.2
starlette==0.37.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.5.0
tifffile==2023.4.12
tiktoken==0.7.0
timeout-decorator==0.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.3.0+cu118
torchaudio==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tqdm==4.64.1
traitlets==5.9.0
transformers==4.41.2
triton==2.3.0
trl==0.8.6
typer==0.12.3
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==2.2.2
uvicorn==0.21.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.2.3
wheel==0.38.4
xxhash==3.2.0
yapf==0.40.2
yarl==1.9.1
zipp==3.19.2
zstandard==0.19.0
